EarthManipulator: improved off-earth motion and transition.
Figure the tangent point on the earth closest to the ray from
eye-to-mouse and use that for dragging calculations. Also, use
vectors from the earth center to the drag point to directly calculate
the rotation. The previous scheme of mapping the pointer motion to a
plane on the earth breaks down near the edges of the earth and doesn't
work well when transitioning off and on the earth's surface.