improvements to the seamless arc distance algorithm
Avoid vector cross product calculations for normal calculations and
left/right predicates; they are too sensitive to floating point error
and can give inconsistent answers when an edge is tested with the end
points in a different order, as is the case when adjacent tiles are
tested. Use trig functions and line intersection tests instead.

In osgearth_seamless, protect the flag display from wireframe mode.