fix EarthManipulator drag mode bugs
When multiple mouse events arrived during a frame -- and before the
camera view matrix was updated -- there was an inconsistency between
the manipulator parameters and the view matrix used for intersection
testing. This was causing very bad stuttering. Also, there was a
problem with the intersection test returning hits on the
osgearth_manip help text, so for now I've added an option to turn it
off.