Fixed scene clamping with annotations
The AnnotationLabelSource wasn't passing the full style, just the
TextSymbol.  Now it's passing in the full style including the
AltitudeSymbol so clamping is turned on.
applyStyle is now called in setMapNode on an AnnotationNode so clamping
can be setup if it needs to be.
Switched the OcclusionCullingCallback to use the
DPLineSegmentIntersector
Renamed style() to getStyle() in LabelNode so it will override the
correct virtual method.