Various patches
Patches to move the implementation of virtual destructors to .cpp files
rather than inline in the headers.  Some versions of GCC have issues
with not generating a vftable at all if a virtual function is not
defined in a cpp file.  This can cause dynamic casts to fail and other
sorts of havok.
Checking for srs validity in ExtrudeGeometryFilter.
Exporting osgEarth::Features::Filter