Incorrect altitude mode in TerrainProfileGraph
The GeoPoint start/end attributes in the TerrainProfileCalculator are
relative, but the TerrainProfileGraph tries to convert the start/end
points toWorld() coordinates, which requires absolute altitudes.

Added overloaded methods to the TerrainProfileCalculator to request the
start/end points in a specific altitude mode, and changed the
TerrainProfileCalculator to use the new TerrainProfileCalculator
interface.

Resolves Issue #532 in the osgearth master.