Added new access methods in Style and Stylesheet
Access methods needed to UI introspection of current applied StyleSheet and Style. This allows our application to inspect current stylesheet applied to a FeatureModel. The end-user can then make changes and reload the selected FeatureModel based on an updated / edited stylesheet.

Note on "void mergeConfig()": method is defined as a public virtual method in derived classes, but in the base class, it was defined as protected and non-virtual.