Cache wms, wfs, and tms metadata per service
If a system has intermittent network issue (e.g. like at an IBC booth), you might not get some layer in an earth file at all if network connection breaks when metadata for layers are read (e.g. WFS Capabilities), event though all tiles are cached locally.

This commit is a hackish solution that caches WFS/WMS/TMS metacast data to disc when read successfully. If then at a later load of an earth file, the network connection is broken at the time when metadata is read, the cached metadata is used.

We would like to see these changes evolve into a propper feature, so that:
* metadata caching can be configured (Enable/disable, age-threshhold, etc.)
* solution is made platform indepedent and consistent with osgEarth data cache config.