VRV Defect 62157 - Extruded and Point replace buildings showing through clouds when alpha to coverage is disabled. Email about the issue:
Hey Glenn,

In October you made a change to ModelLayer.cpp which puts all model layers in the transparent bin. This causes the buildings on alamona to show through the clouds in VR-Vantage since they are then being rendered after the clouds.

The change was part of git rev: 958707d33cca50b4f6cdb80ac46a77a1bc734156 Updates to fix render order issues with splatlayer and other non-image tile layers

Does it make sense to make that change configurable through the layer options? Or can that line be moved to a layer that is more specific to the problem it is fixing?

Thanks,

-Tim

Glenn Waldron
11:55 AM (7 minutes ago)

to me, Jason, VRV
Tim, I will just remove it. No idea why it's there.

Glenn Waldron / Pelican Mapping