Merge from osgEarth master commit:988e7f72e6274c45f368bb6616416b1797464e86 rex: fix mishandling of ProgressCallback cancelation state, which could cause a partially built tile when using more than one elevation layer
This fixed the holes in alamona.mtf


Explanation:

osgEarth has the ability cancel a tile request mid-stream if the engine determines that it no longer needs the tile. This can happen when the camera is moving, and a tile comes into play, and then goes back out of play because of a moving frustum. In the simple case where you have a single elevation layer, that layer will either detect cancelation of not, and you will either get a NULL tile (which is discarded) or a full tile (which is used but not really needed; no harm done).

In the case of multiple elevation layers, cancellation flag might be set between the processing of layers. This is what happened here; the low-resolution elevation layer initialized OK, but the cancellation event occurred before the second (hi-res) layer could initialize. The engine was only able to use height data from the lo-res layer, and you got lo-res tiles.

Fix: I added some checks in the rex engine that were missing, i.e., to do some final checks on the request to make sure it wasn't cancelled along the way.