From Glenn: If the local data source isn't there it fails. That was part of the work to add better error reporting to the Layer system. It also clarifies the intent of the cache which is to avoid costly reprojection/mosaicing of local data or re-fetching of data over the network.
If you don't like that intent, you can remove the "!tileSourceStatus.isError()" clause in your fork and it should fall back on a (maybe existing) cache if the local source file is missing.

This will prevent we read the cache from a gdal driver. So we remove that line to make the gdal cache working.