Moved PrimiveIntersector::Intersection copy ctor (#1459)
The copy constructor for PrimitiveIntersector::Intersection was defined in
the cpp file while the rest of the struct was defined inline in the
header.  This meant that you couldn't copy the struct without from your
own project without getting linker errors.  I simply moved the copy ctor
to be inline like the rest of the struct.