Fix trench effect in mesh edits
The trick is to get the Z coordinates of the cut-in line from the
existing mesh, not the cutting line; its origin may be well outside
of the relevent tile and not correct at all. There may be times when
we want to use the Z coordinate from the cutting features, but not
yet.