Fix target names in osgEarthConfig.cmake.in
The config file is missing a definition for the osgEarth library, which causes a configuration error upon attempting to set properties of the target.

The variable and target names are also not properly capitalized. The exported library name is "osgEarth", but the variables names used in osgEarthConfig.cmake.in prior to this commit are the lowercase "osgearth". 

Related to this capitalization problem is the name of the path provided to find_path(). The path is fully capitalized, which does not match the actual file path. This can cause some systems to fail to find the correct path.